+++ tests/core/tests/admin_integration/test_import.py
@@ -14,7 +14,7 @@ from django.test.testcases import TestCase, TransactionTestCase
 from django.test.utils import override_settings
 from django.utils.translation import gettext_lazy as _
 
-from import_export.admin import ExportActionModelAdmin, ExportMixin
+from import_export.admin import ExportMixin
 from import_export.formats import base_formats
 
 
@@ -433,22 +433,6 @@ class ImportAdminIntegrationTest(AdminTestMixin, TestCase):
     def test_get_context_data_returns_empty_dict(self):
         m = ExportMixin()
         self.assertEqual(dict(), m.get_context_data())
-
-    def test_media_attribute(self):
-        """
-        Test that the 'media' attribute of the ModelAdmin class is overridden to include
-        the project-specific js file.
-        """
-        mock_model = mock.MagicMock()
-        mock_site = mock.MagicMock()
-
-        class TestExportActionModelAdmin(ExportActionModelAdmin):
-            def __init__(self):
-                super().__init__(mock_model, mock_site)
-
-        m = TestExportActionModelAdmin()
-        target_media = m.media
-        self.assertEqual("import_export/action_formats.js", target_media._js[-1])
 
     @patch("import_export.admin.logger")
     def test_issue_1521_change_list_template_as_property(self, mock_logger):