+++ docs/changelog.rst
@@ -21,6 +21,7 @@ Please refer to :doc:`release notes<release_notes>`.
 - Deprecation of ``ExportViewFormMixin`` (#1666)
 - Added :meth:`~import_export.resources.Resource.do_instance_save` helper method (#1668)
 - Refactored test_admin_integration: split into smaller test modules (#1662)
+- Support multiple Resources for export (#1671)
 - Refactored test_resources: split into smaller test modules (#1672)
 - Fix deprecated ``log_action`` method (#1673)
 - Enable defining Resource model as a string (#1669)